Added Alien Structures
This is the first mod added that actually has a conflict.  The conflict
existed in FULL.MBIN and FULLSAFE.MBIN.  For this conflict it was super
easy.  The existing mods had their new models loaded under the Landmarks
section while Alien Structures loads them under Objects.  Adding the
stuff under Landmarks to the new FULL and FULLSAFE files allows the two
mods to work together.