Added Crystal Interactions
Added a new feature: Crystal Interactions.  You can interact with large
crystals instead of destroying them for a chance at a large amount of
any of the 4 types of rare resources!  Beware though, you may also
receive damage in the process.  This ability will eventually require a
custom technology (similar to the Haz Mat Gauntlets) that is given via a
minor quest.