First Commit of City Builder
Can build a teleporter to take you back to base, an observatory, and a
hologram (not working).  Also custom build menu categories work.  This
is proof of concept.  Will need to add in all the other buildings and
determine if snap nodes work on buildings that aren't made to be built
but this is a good step in the right direction!